A · module reorg: read/* → ceg/* topic namespaces (façade preserves API)
Pure mechanical move + re-export per FSD V4.0 §3 (Data Access Surface).
No behaviour change: every read primitive type rehomes under topic-named
src/ceg/ namespaces; src/read/mod.rs becomes a thin `pub use crate::ceg::*`
façade so existing crate::read::* / ciris_persist::read::* paths still
resolve. The façade is removed in a LATER v4.0 commit (FSD §3.3).

- read/trace.rs       -> ceg/list/traces.rs (TraceSummary/Detail/...) +
                        ceg/aggregates/scoring.rs (Divergence/Drift/
                        OverrideRate/HashChainGap)
- read/scoring.rs     -> ceg/aggregates/scoring.rs
- read/llm.rs         -> ceg/list/llm.rs (filter/cursor/page) +
                        ceg/aggregates/llm.rs (cost rollups)
- read/scrub.rs       -> ceg/aggregates/scrub.rs
- read/corpus.rs      -> ceg/aggregates/corpus.rs
- read/federation.rs  -> ceg/list/federation.rs
- read/task.rs        -> ceg/list/tasks.rs
- read/types.rs       -> ceg/types/{window,cursor,filter}.rs
- ReadEngine + Error  -> ceg/mod.rs (flat re-exports preserve every name)

Placeholder topic homes staked out for later commits: ceg/{cohort_scope,
identity,family,community,structural_invisibility,streaming}.

lib.rs gains pub mod ceg; prelude re-exports follow to crate::ceg::* with
identical exported names. Internal call sites resolve through the façade.

cargo check (postgres+server+pyo3+sqlite) green; 569 lib tests pass under
sqlite incl. every read-engine re_* test.
