+++ src/ceg/aggregates/scoring.rs
@@ -1,5 +1,11 @@
-//! Section E — Scoring factor aggregates (CIRIS Capacity Score
-//! primitives).
+//! Section E + F — Scoring factor aggregates (CIRIS Capacity Score
+//! primitives) and Coherence Ratchet divergence/drift/gap rows.
+//!
+//! Moved + merged in v4.0 (FSD §3.3): the section-E scoring shapes from
+//! `src/read/scoring.rs` and the section-F Coherence-Ratchet input rows
+//! (`DivergenceRow`, `TemporalDriftRow`, `OverrideRateRow`,
+//! `HashChainGap`) that previously lived in `src/read/trace.rs` are
+//! co-located here under the aggregates namespace.
 //!
 //! `api/scoring.py` runs N agents × M factors × window each pass; today
 //! that's raw SQL. The clean substrate surface is one bundled aggregate
@@ -7,7 +13,7 @@
 //! in one DB round-trip — plus granular sub-primitives the bundled one
 //! composes from, so analysts can ask narrower questions.
 //!
-//! This module defines the typed shapes; the [`super::ReadEngine`]
+//! This module defines the typed shapes; the [`crate::ceg::ReadEngine`]
 //! trait carries the methods; the Postgres backend implements them.
 //!
 //! ## Capacity Score factor mapping
@@ -31,7 +37,9 @@
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
-use super::types::TimeWindow;
+use crate::ceg::types::{DeviationMetric, TimeWindow};
+
+// ─── Section E — Scoring factor aggregates ─────────────────────────
 
 /// One bundled aggregate covering every factor input for a single
 /// agent + window. Lens consumes this to compute the Capacity Score
@@ -126,7 +134,7 @@ pub struct CoherencePoint {
 }
 
 /// Granular audit-chain aggregate. Returned by
-/// [`super::ReadEngine::aggregate_audit_chain`].
+/// [`crate::ceg::ReadEngine::aggregate_audit_chain`].
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct AuditChainAggregate {
     /// Total audit entries observed in the filter window.
@@ -137,4 +145,82 @@ pub struct AuditChainAggregate {
     pub audit_hashed: i64,
     /// Detected gaps (count of contiguous-sequence breaks).
     pub gap_count: i64,
+}
+
+// ─── Section F — Coherence Ratchet inputs ──────────────────────────
+
+/// One agent's divergence from the deployment-domain peer mean over
+/// a window. Lens computes detection (clustering, threshold) from
+/// these inputs; persist provides the windowed peer-mean reference.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct DivergenceRow {
+    /// Agent identity.
+    pub agent_id_hash: String,
+    /// Optional human-readable agent name.
+    pub agent_name: Option<String>,
+    /// Z-score against the domain peer mean.
+    pub z_score: f64,
+    /// Which metric drove the divergence.
+    pub deviation_metric: DeviationMetric,
+    /// Trace count contributing to this z-score (lens-side k-anon
+    /// filtering — see AV-43 in THREAT_MODEL.md).
+    pub sample_count: i64,
+}
+
+/// Drift between two windows for a single agent on a single metric.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct TemporalDriftRow {
+    /// The metric this row reports.
+    pub deviation_metric: DeviationMetric,
+    /// Baseline (earlier) window.
+    pub baseline_window: TimeWindow,
+    /// Comparison (later) window.
+    pub comparison_window: TimeWindow,
+    /// `comparison_mean - baseline_mean`.
+    pub mean_shift: f64,
+    /// `comparison_var / baseline_var`.
+    pub variance_ratio: f64,
+    /// Significance metric (z-score under normal-mean approximation).
+    /// Lens applies its own p-value mapping.
+    pub significance: f64,
+}
+
+/// One detected gap in the agent's audit-chain sequence number
+/// timeline. `gap_start_seq` is the last-seen seq before the gap;
+/// `gap_end_seq` is the first-seen seq after.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct HashChainGap {
+    /// Agent identity.
+    pub agent_id_hash: String,
+    /// Last sequence number seen before the gap.
+    pub gap_start_seq: i64,
+    /// First sequence number seen after the gap.
+    pub gap_end_seq: i64,
+    /// Wall-clock of the last pre-gap entry.
+    pub gap_start_ts: DateTime<Utc>,
+    /// Wall-clock of the first post-gap entry.
+    pub gap_end_ts: DateTime<Utc>,
+}
+
+/// One agent's conscience-override rate over a window, with the
+/// deployment-domain average for ratio computation.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct OverrideRateRow {
+    /// Agent identity.
+    pub agent_id_hash: String,
+    /// Optional human-readable agent name.
+    pub agent_name: Option<String>,
+    /// Deployment domain (echoed for caller convenience).
+    pub deployment_domain: Option<String>,
+    /// Number of conscience overrides observed in the window.
+    pub override_count: i64,
+    /// Total trace count in the window (denominator).
+    pub trace_count: i64,
+    /// `override_count / trace_count`. `0.0` when `trace_count == 0`.
+    pub override_rate: f64,
+    /// Average override rate across all agents in the same domain.
+    pub domain_avg_rate: f64,
+    /// `override_rate / domain_avg_rate`. `1.0` when both are equal;
+    /// >1.0 means this agent overrides more than peers.
+    pub multiple_of_domain_avg: f64,
 }
