+++ src/ceg/cohort_scope/mod.rs
@@ -1,0 +1,7 @@
+//! CEG 0.4 cohort_scope wire-format; admission-gate semantics added
+//! 0.10 §10.1.4.
+//!
+//! Placeholder home (v4.0 FSD §3.1). The `classification.rs`
+//! (`suppresses_holds_bytes`, closed-set validation) and `predicate.rs`
+//! (the §4.3 SQL predicate) modules land in a later v4.0 commit; this
+//! commit only stakes out the topic namespace.
