+++ src/ceg/community/mod.rs
@@ -1,0 +1,7 @@
+//! CEG 0.10 §8.1.13.3 community reads (V060).
+//!
+//! Placeholder home (v4.0 FSD §3.1). The community-read primitives land
+//! in a later v4.0 commit alongside the V060 `federation_communities`
+//! substrate; this commit only stakes out the topic namespace. These
+//! reads *consume* the `crate::federation` substrate; they do not
+//! replace it.
