+++ src/ceg/family/mod.rs
@@ -1,0 +1,6 @@
+//! CEG 0.7 §5.6.8.9 family reads.
+//!
+//! Placeholder home (v4.0 FSD §3.1). The family-read primitives land in
+//! a later v4.0 commit; this commit only stakes out the topic
+//! namespace. These reads *consume* the `crate::federation` substrate
+//! (`federation_families`); they do not replace it.
