+++ src/ceg/identity/mod.rs
@@ -1,0 +1,6 @@
+//! CEG 0.7 §5.6.8.8 identity_occurrence reads.
+//!
+//! Placeholder home (v4.0 FSD §3.1). The `occurrence.rs` reader lands
+//! in a later v4.0 commit; this commit only stakes out the topic
+//! namespace. These reads *consume* the `crate::federation` substrate
+//! (`federation_identity_occurrences`); they do not replace it.
