+++ src/ceg/list/traces.rs
@@ -1,5 +1,10 @@
-//! Section A + B + F types — trace listing, trace detail, Coherence
-//! Ratchet inputs.
+//! Section A + B types — trace listing + trace detail.
+//!
+//! CEG 0.4 trace wire-format; lens / lens-core / sovereign-agent read
+//! surface (CIRISPersist#23 §A/§B). Moved from `src/read/trace.rs` in
+//! v4.0 (FSD §3.3) — the section-F Coherence-Ratchet scoring rows that
+//! previously shared `read/trace.rs` now live in
+//! `src/ceg/aggregates/scoring.rs`.
 //!
 //! These structs cross both the rlib path (Rust-public) and the PyO3
 //! path (typed dicts on the Python side; field shape identical). Wire-
@@ -8,9 +13,8 @@
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
+use crate::ceg::types::TraceCursor;
 use crate::schema::{ReasoningEventType, TraceLevel};
-
-use super::types::{DeviationMetric, TimeWindow, TraceCursor};
 
 // ─── Section A — Trace listing ─────────────────────────────────────
 
@@ -176,82 +180,4 @@ pub struct TraceEnvelopeRefs {
     pub scrub_timestamp: Option<DateTime<Utc>>,
     /// True after the scrubber pass ran.
     pub pii_scrubbed: bool,
-}
-
-// ─── Section F — Coherence Ratchet inputs ──────────────────────────
-
-/// One agent's divergence from the deployment-domain peer mean over
-/// a window. Lens computes detection (clustering, threshold) from
-/// these inputs; persist provides the windowed peer-mean reference.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct DivergenceRow {
-    /// Agent identity.
-    pub agent_id_hash: String,
-    /// Optional human-readable agent name.
-    pub agent_name: Option<String>,
-    /// Z-score against the domain peer mean.
-    pub z_score: f64,
-    /// Which metric drove the divergence.
-    pub deviation_metric: DeviationMetric,
-    /// Trace count contributing to this z-score (lens-side k-anon
-    /// filtering — see AV-43 in THREAT_MODEL.md).
-    pub sample_count: i64,
-}
-
-/// Drift between two windows for a single agent on a single metric.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct TemporalDriftRow {
-    /// The metric this row reports.
-    pub deviation_metric: DeviationMetric,
-    /// Baseline (earlier) window.
-    pub baseline_window: TimeWindow,
-    /// Comparison (later) window.
-    pub comparison_window: TimeWindow,
-    /// `comparison_mean - baseline_mean`.
-    pub mean_shift: f64,
-    /// `comparison_var / baseline_var`.
-    pub variance_ratio: f64,
-    /// Significance metric (z-score under normal-mean approximation).
-    /// Lens applies its own p-value mapping.
-    pub significance: f64,
-}
-
-/// One detected gap in the agent's audit-chain sequence number
-/// timeline. `gap_start_seq` is the last-seen seq before the gap;
-/// `gap_end_seq` is the first-seen seq after.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct HashChainGap {
-    /// Agent identity.
-    pub agent_id_hash: String,
-    /// Last sequence number seen before the gap.
-    pub gap_start_seq: i64,
-    /// First sequence number seen after the gap.
-    pub gap_end_seq: i64,
-    /// Wall-clock of the last pre-gap entry.
-    pub gap_start_ts: DateTime<Utc>,
-    /// Wall-clock of the first post-gap entry.
-    pub gap_end_ts: DateTime<Utc>,
-}
-
-/// One agent's conscience-override rate over a window, with the
-/// deployment-domain average for ratio computation.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct OverrideRateRow {
-    /// Agent identity.
-    pub agent_id_hash: String,
-    /// Optional human-readable agent name.
-    pub agent_name: Option<String>,
-    /// Deployment domain (echoed for caller convenience).
-    pub deployment_domain: Option<String>,
-    /// Number of conscience overrides observed in the window.
-    pub override_count: i64,
-    /// Total trace count in the window (denominator).
-    pub trace_count: i64,
-    /// `override_count / trace_count`. `0.0` when `trace_count == 0`.
-    pub override_rate: f64,
-    /// Average override rate across all agents in the same domain.
-    pub domain_avg_rate: f64,
-    /// `override_rate / domain_avg_rate`. `1.0` when both are equal;
-    /// >1.0 means this agent overrides more than peers.
-    pub multiple_of_domain_avg: f64,
 }
