+++ src/ceg/streaming/mod.rs
@@ -1,0 +1,5 @@
+//! CEG 0.10 streaming attestation reads.
+//!
+//! Placeholder home (v4.0 FSD §3.1). The streaming-read primitives land
+//! in a later v4.0 commit; this commit only stakes out the topic
+//! namespace.
