+++ src/ceg/structural_invisibility/mod.rs
@@ -1,0 +1,5 @@
+//! CEG 0.10 §10.1.4 — the read-side structural-invisibility gate.
+//!
+//! Placeholder home (v4.0 FSD §3.1). The read-side cohort_scope gate
+//! lands in a later v4.0 commit; this commit only stakes out the topic
+//! namespace.
